move autocomplete locations to its own component file
