+++ src/components/AutocompleteLocations/index.ts
@@ -1,0 +1,5 @@
+import AutocompleteLocations from './AutocompleteLocations';
+import { isCounty, isState, belongsToState } from './utils';
+
+export default AutocompleteLocations;
+export { isCounty, isState, belongsToState };
